// fantastiscotch/src/types/FlavorNames.ts
enum FlavorNames {
  finish = "Finish",
  acidic = "Acidic",

  smoke = "Smoke",
  malty = "Malty",

  floral = "Floral",
  balance = "Balance",

  salty = "Salty",
  peat = "Peat",

  herbal = "Herbal",
  citrus = "Citrus",

  body = "Body",
  sweet = "Sweet",

  toffee = "Toffee",
  spicy = "Spicy",

  darkFruit = "Dark Fruit"
}

export default FlavorNames;